Added Type 1a test
Added a test that determines weather a given vertex, or tangent point of a circle, tested for a concave support can be a concave support.
Note: this method will not work, if a vertex can be placed in a 1a type zone, above the line from s2-r.  Fixed next commit.